+++ tests/setup-e2e.js
@@ -15,7 +15,12 @@ import {
   generateScreenshotFilePath,
   getTestNameFromTestFile,
   saveScreenshot,
+  waitAndClick,
+  waitAndGetText,
+  waitAndSetValue,
+  skippablePromise,
 } from './common/e2e/steps/helpers';
+import { DEFAULT_TIMEOUT } from './common/e2e/steps/config';
 import { setNewsFeedIsOpen, resetTestNews } from './news/e2e/steps/newsfeed-steps';
 import { refreshClient } from './nodes/e2e/steps/helpers';
 import { TEST } from '../source/common/types/environment.types';
@@ -25,7 +30,6 @@ import type { Daedalus } from './types';
 
 declare var daedalus: Daedalus;
 const context = {};
-const DEFAULT_TIMEOUT = 20000;
 let scenariosCount = 0;
 
 const printMainProcessLogs = () =>
@@ -103,9 +107,11 @@ Before({ tags: '@e2e', timeout: DEFAULT_TIMEOUT * 2 }, async function(testCase) 
   await this.client.executeAsync(done => {
     const resetBackend = () => {
       if (daedalus.stores.networkStatus.isConnected) {
+        daedalus.api.ada.resetTestOverrides();
         daedalus.api.ada
           .testReset()
           .then(daedalus.api.localStorage.reset)
+          .then(daedalus.stores.wallets.refreshWalletsData())
           .then(done)
           .catch(error => {
             throw error;
@@ -140,12 +146,14 @@ Before({ tags: '@newsfeed' }, function() {
   resetTestNews(this.client);
 });
 
+
 // adds waitAndClick method to webdriver
-Before(function() {
-  this.waitAndClick = async (selector, ...waitArgs) => {
-    await this.client.waitForVisible(selector, ...waitArgs);
-    return this.client.click(selector);
-  };
+Before(function(testCase) {
+  const { name } = testCase.pickle;
+  this.skippablePromise = skippablePromise.bind(this, name);
+  this.waitAndClick = waitAndClick.bind(this);
+  this.waitAndGetText = waitAndGetText.bind(this);
+  this.waitAndSetValue = waitAndSetValue.bind(this);
 });
 
 // ads intl method to webdriver